Fixing checkpoints
We need to use a null value when checking for difficulty transitions
as an indicator that we're syncing from a checkpoint.  This is the
only way to tell in this part of the code.  If we're not using a checkpoint,
the transition base (back an interval) will be non-null.  In the case
of the genesis block, we use interval - 1.